feature(generator): Moved all busines logic to the custom hook
